# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api rest Documentation for Tasks",
        "description": "Api Rest for educational purposes",
        "version": "0.0.1"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User authentication",
                "description": "Get a JWT via given credentials.",
                "operationId": "d33fa59eac86d58bb7a93b3ff8a3b7f1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User authenticate"
                    },
                    "401": {
                        "description": "User Unauthorized"
                    },
                    "default": {
                        "description": "An error occurred"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User logout",
                "description": "Log the user out (Invalidate the token).",
                "operationId": "29573ceac77e01db83ed67bd710fc374",
                "responses": {
                    "200": {
                        "description": "User logout"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "default": {
                        "description": "An error occurred"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/tasks": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Show task list",
                "description": "Display a listing of the resource.",
                "operationId": "471d938a3bfaeb6fbd981d7c3e2728b0",
                "responses": {
                    "200": {
                        "description": "Show all tasks"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "default": {
                        "description": "An error occurred"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Show task info",
                "description": "Display the specified resource.",
                "operationId": "90b3cd4d2d4b416dc989d2055dbed06c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter to search the task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "Enter a task id number",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show task info"
                    },
                    "404": {
                        "description": "The task has not been found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "default": {
                        "description": "An error occurred"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}